+++ example/demo.f90
@@ -1,6 +1,6 @@
 ! The MIT License (MIT)
 !
-! Copyright (c) 2023 Vincent Magnin
+! Copyright (c) 2023-2024 Vincent Magnin
 !
 ! Permission is hereby granted, free of charge, to any person obtaining a copy
 ! of this software and associated documentation files (the "Software"), to deal
@@ -21,11 +21,12 @@
 ! SOFTWARE.
 !-------------------------------------------------------------------------------
 ! Contributed by vmagnin: 2023-09-26
-! Last modification: gha3mi 2024-01-28, vmagnin 2024-02-15
+! Last modification: gha3mi 2024-01-28, vmagnin 2024-02-22
 !-------------------------------------------------------------------------------
 
 program demo
     use forcolormap, only: Colormap, colormaps_list, wp
+    use forcolormap_utils, only: test_colormap
     implicit none
 
     integer :: i
@@ -43,12 +44,12 @@ program demo
         255,   255,   255 ], &
         shape(my_colormap), order = [2, 1] )
 
-    ! Let's create PPM files for each built-in colormap.
-    ! The built-in z=f(x,y) test function is in the [0, 2] range:
+    !> We create PPM files (binary encoded by default) for each built-in colormap.
+    !> The built-in z=f(x,y) test function is in the [0, 2] range:
     do i = 1, size(colormaps_list)
         call cmap%set(trim(colormaps_list(i)), 0.0_wp, 2.0_wp)
         call cmap%colorbar(trim(colormaps_list(i))//'_colorbar')
-        call test_colormap(cmap, trim(colormaps_list(i))//'_test', encoding='binary')
+        call test_colormap(cmap, trim(colormaps_list(i))//'_test')
         print '("Colormap ", A30, " has ", I0, " levels")', trim(cmap%get_name()), cmap%get_levels()
     end do
 
@@ -56,53 +57,17 @@ program demo
     call cmap%set("cubehelix", 0.0_wp, 2.0_wp, 1024, [0.5_wp, -1.0_wp, 1.0_wp, 1.0_wp])
     ! We change the name for the output test files:
     call cmap%colorbar('cubehelix_customized_colorbar')
-    call test_colormap(cmap, 'cubehelix_customized_test', encoding='binary')
+    call test_colormap(cmap, 'cubehelix_customized_test')
 
     ! You can create your own colormap defined in an array:
     call custom_cmap%create('discrete', 0.0_wp, 2.0_wp, my_colormap)
     call custom_cmap%colorbar('discrete_colorbar')
-    call test_colormap(custom_cmap, 'discrete_test', encoding='binary')
+    call test_colormap(custom_cmap, 'discrete_test')
 
     ! Or you can download it from a .txt file:
     call custom_cmap%load("test_map_to_load.txt", 0.0_wp, 2.0_wp)
     call custom_cmap%colorbar('a_loaded_colorbar')
-    call test_colormap(custom_cmap, 'a_loaded_colormap_test', encoding='binary')
+    call test_colormap(custom_cmap, 'a_loaded_colormap_test')
     call custom_cmap%print()
 
-    contains
-
-    subroutine test_colormap(self, filename, encoding)
-        use forimage, only: format_pnm
-        type(Colormap), intent(inout) :: self
-        character(*), intent(in) :: filename
-        integer :: k, j     ! Pixbuffer coordinates
-        integer, parameter :: pixwidth  = 600
-        integer, parameter :: pixheight = 600
-        integer, dimension(pixheight,pixwidth*3) :: rgb_image
-        integer  :: red, green, blue
-        real(wp) :: z
-        type(format_pnm) :: ppm
-        character(*), intent(in) :: encoding
-
-        do k = 0, pixwidth-1
-            do j = 0, pixheight-1
-                ! Computing a z=f(x,y) function:
-                z = 1.0_wp + sin(k*j/10000.0_wp) * cos(j/100.0_wp)
-                ! The corresponding RGB values in our colormap:
-                call self%compute_RGB(z, red, green, blue)
-                rgb_image(pixheight-j, 3*(k+1)-2) = red
-                rgb_image(pixheight-j, 3*(k+1)-1) = green
-                rgb_image(pixheight-j, 3*(k+1))   = blue  
-            end do
-        end do
-
-        call ppm%set_pnm(encoding    = encoding,&
-                        file_format = 'ppm',&
-                        width       = pixwidth,&
-                        height      = pixheight,&
-                        max_color   = 255,&
-                        comment     = 'comment',&
-                        pixels      = rgb_image)
-        call ppm%export_pnm(filename)
-    end subroutine test_colormap
 end program demo