+++ CHANGELOG.md
@@ -29,15 +29,16 @@ All notable changes to the gtk-fortran project are documented in this file. The 
 * A `reverse` boolean option was added to the methods `set`, `create`, and `load` to reverse a colormap.
   * An `example/demo_reverse.f90` example.
 * The "magma", "inferno","plasma", "viridis" matplotlib colormaps in a `matplotlib_colormaps` module.
+* A `src/forcolormap_utils.f90` module with the subroutine `test_colormap()`.
 * A `colormap_parameters` module.
 * A `miscellaneous_colormaps` module.
 * A `colorbar()` type-bound procedure to write a PPM file with the colorbar.
 
 ### Changed
+* Subroutine `test_colormap()`: the `encoding` argument is now optional (binary by default).
 * The colormap `inverted_rainbow` was renamed `inv_rainbow`.
 * `TODO.md` was refactored and renamed `ROADMAP.md`.
 * Code refactoring.
-* Moved `test` subroutine to demo and example1.
 * Renamed `get_current()` to `get_name()`.
 * For writing PPM files, the project [ForImage](https://github.com/gha3mi/forimage) by @gha3mi is now used as a fpm dependency. The related example must now be launched with `fpm run --example example1` and the ForColormap demo by `fpm run --example demo`.
 * `src/colormap_class.f90`: `private` statement is now the default.